Draw: Fix sprite plotting with SpriteExtend 1.43 and later.
In numerous places where a sprite area isn't strictly necessary !Draw was passing in a bogus area address of 0xFF, this is now rejected by SpriteExtend.
Changed to use define UNUSED_SA and set this to an address which is at least in the application slot and word aligned.
Tested briefly - sprites now appear again.

Version 1.19. Tagged as 'Draw-1_19'