BootCmds: Fix to the AddApp command.
If you attempt to AddApp an app which is already within Resources:$.Apps (but not
a subdirectory therein) it would create a stub with circular references, leading
to stack exhaustion. Fixed to take no action in this case (i.e. not to create the
stub).

Untested. Will be required for booting from ResourceFS.

Version 1.40. Tagged as 'BootCmds-1_40'