DrawFile: Ungarble the help and syntax messages.
At some point in internationalisation the help and syntax messages got muddled, so if you did "*RENDER" (ie. a syntax error because no filename is given) you got both the syntax message and then overwritten with the help text (because it has carriage returns in it).
Help change to do <plain english><options><syntax> and
      syntax does <syntax>
like *COPY command.

Version 1.57. Tagged as 'DrawFile-1_57'