HAL_IOMD: Addition of POST checking
Unlike some more recent targets, there are a lot of optional/removeable bits on IOMD platforms which may stop the machine booting, this checks for
* VRAM (non fatal if missing)
* RAM
* ROM checksum
* Video timing
* IOMD
in the event of a failure the fault code is flashed on the floppy drive LED (pattern decoder in TRM appendix A).

Video.s: the POST code highlighted that the VIDCTAB had the wrong VCO reference clock on Medusa - this is now corrected for a 60Hz mode.
IOMD.hdr: share the address line bits
Tested on Risc PC + A7000, the timing emulation is not accurate enough on RPCEmu, so POST is skipped under emulation.

Version 0.18. Tagged as 'IOMD-0_18'