Filer: Swap hot key for something less fat finger prone
When used with 'confirm deletes' off (which is probably a common mode of operation so as to not have to confirm every single file) it's now all too easy to delete whole directories of files if selected.
Shortcut key changed to CTRL-X (to match the build in apps) and pave the way for a future with CTRL-C and CTRL-V ala Windows.

Version 2.31. Tagged as 'Filer-2_31'