FileSwitch: Fix dereference of duff FS control block when resolving badly formed paths
In OS_Find there is a call to TopPath_DoBusinessToPath in DoTheOpen, any V set exits from this are propagated to the user correctly.
Non V set exits (ie. the object type is returned) are passed to DoTheOpen_Common which in turn calls TryToOpenFile. TryToOpenFile (if fscb is not Nowt) dereferences it to get at the info word.
A problem exists where TopPath_DoBusinessToPath fails the policing check - when multipart paths are permitted (such as OS_Find) it swallows the error, steps the path, and eventually exits with an object type 0. However, this code path never sets up the fscb.
So when TryToOpenFile dereferences it (in testing it seemed to have some address in the kernel lingering) depending on your luck the fsinfo_alwaysopen bit might be set. This in turn triggers a forced call to the low level FS entry point, and the entry point is some garbage address leading to an abort.
The minimalist test cases were
 SYS"OS_Find",&41,"MyName","MyPath$"
 SYS"OS_Find",&49,"MyName","MyPath$"
both of which fail the policing step because the incorrect use of the path type bits mean the attempted name becomes MyPath$MyName and you can't have '$' in the middle of a filename. Other more complex examples exist, including having a multipart path where one of the path elements in invalid.
Solved here by setting fscb up to a default (valid) value.
Fixes ticket #287.

Version 2.80. Tagged as 'FileSwitch-2_80'