ConfigLib: Prefer to ask the Kernel for a reset CMOS value
Before looking for a ResetCMOS file, try asking the Kernel, as the mapping between OS version (and hence boot hook directory) isn't as simple as it was in Acorn's time - there can be several wildly differing platforms all from the same OS version.
Part of ticket #316.

Version 0.10. Tagged as 'Configure-0_10'