HAL_IOMD: Run boot time keyboard scan in parallel with RAM clear
Previously the RAM clear would take place (in the kernel startup) then 5s of waiting for the keyboard would occur.
Now, if a keypress is detected the scan is terminated early, or if 2s has elapsed (allowing for the keyboard's POST plus a typematic repeat plus a bit of slack).
This way, if your RAM clear takes 2s or more the boot process is 5s faster.
Tested on a Risc PC with 128MB of RAM.

Version 0.28. Tagged as 'IOMD-0_28'