!Perl: Rebuild of perl binary
Perl suffered from random "out of memory" errors when lots was available for seemingly no reason.
Perl binary rebuilt (in the Disc environment) with cc 5.69, this seems to stop the problem, so chalking this one up to a compiler fault.

Tested by completing a disc image rebuild with no errors.
Fixes ticket #304, for some value of 'fix'.
Tagged as Perl-2013-05-27.