ABRelease: Ammend PlingSystem obey script
The !SysMerge obey script hadn't taken into account that the sprites files for !System need (one off) moving into a subdirectory to match the !Run and !Boot that it installs. This prevents an error when trying to iconsprites a non existant directory.
Tested on RISC OS 4.02.

Version 0.57. Tagged as 'ABRelease-0_57'