DOSFS: Fixes to stop DOSFS erroneously claiming partitions that aren't valid
In particular with ADFS, attempting to format an ATA drive purchased from a PC box shifter would often lead to the machine hanging on startup meaning you can't get as far as running !HForm. Further analysis showed that the problem was caused by DOSFS latching onto the PC partition table and attempting to claim the Service_IdentifyDisc. This would then be followed by trying to load the FAT via FileCore in situations where the FAT to load were bigger than the disc size represented by the disc record FileCore has attached to the disc.

This primarily concerns things with partition tables (like a fixed disc).
Things got off to a bad start, some inaccurate keyboard mashing meant the definition for DOS_bootsector added up to 520 bytes, and this was loaded into a malloc'd 512 byte buffer.
The representation of the disc size (bbsize) was being kept in a signed integer, so the single (rather weak) test of checking the boot block declared size (bbsize) being greater than the partition declared size (winisize) would always be accepted for partitions > 2GB.
Additionally, the bbsize was multiplied by DOSsecsize and performed as a byte compare rather than a sector compare, leading to overflow.
Invariably this check would pass (either due to the sign bit being set, or overflow) and DOSFS would claim the service and go ahead and try to read in an unfeasibly large FAT.

DOSFS is now much more strict about checking
* be sure bbsize can fit in the partition by comparing as unsigned number of sectors, convert to bytes later
* be sure bbsize is non zero
* be sure the FAT isn't bigger than bbsize
* be sure the root directory isn't bigger than bbsize
only then claim the service.

DOSshape.h:
Correct the size of the DOS_bootsector back to 512 bytes.
ADFSshape.h:
Remove unused grubby FileCore constants, especially as we're not likely to encounter any 256B/sector drives!
Add the extra disc record fields to get us up to 'BigDisc' in FileCore parlance.
DOSFS.c:
If the rootsize is zero, have a go at using the FAT32 ones instead. Previously this logic only triggered if the FAT32 extended signature was present, but that signature is for the fields that follow it - the FAT size preceeds.
Only update the return variables if we're confident we're going to claim the service.
Remove dead variable 'notracks'.

Tested on a sample of 4 ATA drives > 10GB from different vendors previously used in PCs. All 4 used to hang when installed, now they do not.

Version 1.08. Tagged as 'DOSFS-1_08'