Filer: Minor filer fixes
When a viewer is in 'sort by size' mode the size of image files was not being considered. So, a dir full of ZIP files (SparkFS not run) would be in size order amongst the other files, but if SparkFS was running the image files would be sorted as dirs of 0 size.
Fix for ticket #400, the *Filer_Run command now defaults to '-noshift' and inspecting the shift key state must be requested (opt in instead of opt out). This solves two key annoyances, running the !Boot sequence and holding shift causes you to enter the desktop with the obey files from the Tasks directory unhelpfully loaded in a text editor, and when setting a task alarm in !Alarm if the alarm happens to go off when you're typing with the shift key the task's !Run file opens in a text editor.

SortDir.s: count image files as files so they're sorted like they were in Filer 2.26
Commands.s: parse new option switch, and consider -noshift the default
Messages/HelpSrc: add the switch to the command syntax

Tested with a task alarm in !Alarm with and without shift held down, and running a !Boot sequence similarly.

Version 2.40. Tagged as 'Filer-2_40'