Gadgets: ScrollList gadget redraw fixes
When _set_colour() is called, remember to redraw the gadget if the colours changed.
When _set_font() is called, remember to redraw the gadget and try to preserve the scroll bar positions.
Save a couple of SWIs if we know the scroll bars must jump in 2 other places.
Remove ScrollLisS, its (never called) function did nothing but return.
Removed duplicate redraw_box, same as redraw_all_box and marked as const.

Reported by Fred Graute, reference
  https://www.riscosopen.org/forum/forums/4/topics/3628
the memory leak aspects of the report have been transferred to ticket #414 for later consideration.
Tested with !ExifEdit and some special versions from Fred with the redraw workarounds removed.

Version 0.39. Tagged as 'Gadgets-0_39'