FileSwitch: Erase freshly allocated file buffers.
If the buffers are not zeroed the previous contents of the memory
may get written to the file.

A spy may then read said memory contents by examining the disc,
or by reading the files before they are closed (or if the files
are never closed because RISC OS is uncleanly terminated).