SATADriver: Don't punish controllers that report DET_DEV_NE
For those controllers relying on the retry loop to get the top level port to detect, the change in SATADriver 0.05 made the retries take 25s.
Roll back the COMRESET_RECOVERY_TIME to 50cs like it was in revision 1.1.1.1 of op.c.
Then, pull out the subsequent PM rescan code into a function, and wrap it in a retry loop so the net result is the same as SATADriver 0.05 achieved. However, it is possible to escape the loop early with this arrangement, rather than just idling.

Version 0.06. Tagged as 'SATADriver-0_06'