Linux Internet: New Module
This module implements the Socket SWIs using the corresponding
Linux system cals.

In it's current state it suffcient to browse the RISC OS Open
website with Netsurf 3.7 if the IP address is entered into the
hosts file.

Many more SWIs ned to be implemented, and currently DNS does not
work.