FileSwitch: Allow pass-though of object type 3 (image file)
If an filing system presents FileSwitch with an object of
type 3 (image) FileSwitch will not attach an image filing system.

This is done using an extra bit of the object type to internally
indicate an object is an image and mask it off before return to
callers.

Tested with IXFS in Linux port.