Wimp: Fix for bad error pointer deleting non existent sprite Doing *WimpKillSprite _not_there_  had the following behaviour
Priority   | Sprite exists  Sprite doesn't  Sprite is ROM
Lo then Hi | Deletes        Bad error ptr   Silence
Hi then Lo | Deletes        Silence         Silence

The bad error pointer was caused by an early return if the high priority pool is the ROM sprites (because the earlier failure to find it in the low priority pools set V, and the early return used TEQ, so the error pointer in R0 was the SpriteOp reason code).
Proceed to try to look up the sprite, even if there's no hope of deleting it, in order to get a valid error message.

Version 5.61. Tagged as 'Wimp-5_61'