BASIC: When assembly ends call OS_SynchroniseCodeAreas based on O% when assembling to O%.
BASIC 1.75 to switched calling syncing the range old P% to new P% at the
end of assembly instead of all memory. This is incorrect when assembling
to O% and causes random crashes when building FSLock 1.22.

Fix by storing both P% and O% at the start of assembly and choose
the appropriate range based on the OPT value.

Also move the now larger routine to avoid ADR out of range error.