OmniLanManFS: Makefile simplification
Use the INSTRES_FILES specifier to pick up the resource files and generate the Messages rather than having a local static dependency.

Version 2.62. Not tagged

OmniLanManFS: Build fix

Ensure that when built for ROM, the RES_OBJ is created, as LanManFS is a little unusual in carrying its own 3rd party resources.

Version 2.62. Not tagged