BuildSys: Support application and module targets with differing compiler flags
Typically useful for libraries where you have both types of target, and you
want the application build to contain function names (for meaningful
backtraces) but the module build not to (to save ROM space). This is
achieved as follows:

* CAPPFLAGS and CMODFLAGS contain additional flags to be passed to application
  and module builds, respectively

* C_FNAMES is initialised to a compiler-agnostic build switch to enable
  function names in the binary

Therefore a master makefile will often use the line
CAPPFLAGS = ${C_FNAMES}
although CAPPFLAGS and CMODFLAGS are not limited to this usage.