Kernel: Permit special field in paths passed to OS_CLI
OS_CLI permits commands to be prefixed with a temporary filing
system, for example "ADFS:Format"

It also passes unrecognised to FileSwitch to be executed, so
"ADFS::HardDisc4.$.!Boot" first sets the temporary filing to
ADFS then executes ":HardDisc4.$.!Boot".

This does not work if the path contains a special field,
previously this resulting in an error as the special filed
would be lost. Instead if a filing system prefix with special
field is found switch to the Temporary filing system and
treat the whole command as a path to execute. For example
passing "IFXS#R:$.HardDisc4.!Boot" sets the temporary filing
system to "IXFS" then executes "IFXS#R:$.HardDisc4.!Boot".