Kernel: Always display error on boot
Previously an error from returned FSControl_BootupFS would not be
displayed by the kernel with the intention that it would displayed
by the Desktop module, unless XOS_SetVarVal returns an error.

This had result that the error would be hidden if the desktop does
not start for any reason. Fix by unconditionally displying the error.