Draw: [445] Fix for failing to transform text areas properly
The calculation in scale_textC() was using the origin of the group rather than the origin of the text column, and hence when a group of text areas was transformed (xscale/yscale/magnify) all the text got clumped up in the top left corner.
Delete scale_textcolumn() since Draw object 10 can only occur nested inside a text area, there's no need to have a handler for it.

Version 1.39. Tagged as 'Draw-1_39'