BuildSys: Allow cross-compile C++FLAGS to be inherited from environment
Occasionally, it is useful to be able to set C++FLAGS from the calling
environment, for example when building a 32-bit x86 target from a 64-bit
x86-64 host. This works fine for CFLAGS and LDFLAGS, but C++FLAGS is not a
valid Posix environment variable name. Import setting from CXXFLAGS instead
when cross-compiling.