OSLib: Support cross-compilation
While this component is stored in pre-compiled form, we need the header
filenames to be converted to postfix-extension form during the export phase
so that compilers and other code analysis tools will find them. Because this
is a somewhat unusual component, this is achieved with a custom makefile for
cross-compilation use. Support is also included for converting the object
files to ELF format when exporting, if `${TOOLCHAIN}` is `GNU`.

Enable GitLab CI (native `Makefile` given partial GNU make compatibility to
facilite some of the CI jobs, and as a side-effect will enable correct syntax
colouring in GitLab).


Tagged as 'OSLib-6_90-2'