Debugger: Remove CSDB architecture version warnings
Although CSDB was only codified at around the time that ARMv8.4 was
introduced, there aren't any architecture version restrictions placed on
it in the ARMv8 ARM, and the encoding was even added to a new edition of
the ARMv7 ARM, also without any architecture version restrictions.

For ordinary instructions this would feel a lot like a documentation
error, but since CSDB is a barrier instruction which fits inside a
NOP-like address space (MSR which doesn't write to any PSR fields), and
it was introduced as part of Arm's Spectre/Meltdown mitigations, it's
likely that Arm did their homework and confirmed that the encoding has
the correct behaviour on all (licensed) implementations.

The ARMv7 ARM only goes back to ARMv4, but it's likely that older CPUs
also implement CSDB correctly, by virtue of the fact that they'll be too
old to be doing any speculative execution.

With the above in mind, it should be safe to remove our "ARMv8.4 or
later" warning from the disassembly.

Version 2.08. Tagged as 'Debugger-2_08'