HdrSrc: Only include Hdr:Machine.<Machine> from Hdr:CPU.Arch if not already
The logic behind this is to permit the calling source file to override which
`Machine` type is used, by `GET`ting a different one first. This can be
useful for a component that can only be built for one `Machine` type
(for example, many IOMD-only driver modules).

Version 2.99. Not tagged