FPASC: [515] Fix for loop converting packed decimal infinity/NaNs
On encountering an exponent > &4100 (or > &FE0000 for extended packed) to consider as a NaN or infinity, the emulation ends up in LDFConvert_Packed_NaNInf.
However, this is written as a leaf function and returns with MOV PC,LR when the singular caller of it doesn't expect a return so the LR value is actually still pointing at the last BL around line 1926. This creates an infinite loop.
Escape the loop by popping the true return value which was stacked on entry.

Version 4.38. Tagged as 'FPASC-4_38'