TaskWindow: Don't crash on "%TaskWindow -display -ctrl"
Detail:
s.TaskMan: Alter initial value of PassOnVectors in inittask

This fixes a bug, that if an editor is not currently running then
starting a task window with "%TaskWindow -display -ctrl" from
the new task option in the task manager causes a branch through
zero.

Whilst I have not worked out the exact detail of the crash,
the cause appears to be that PassOnVectors is set to false before
GetParentTask is called. In GetParentTask, Wimp_StartTask is
called, which emits control characters. This causes
GetParentTask to be re-entered (this time via MyWrchV rather
than the start-up -display code). It gets hazy what happens
after to branch through zero, but changing it to set
PassOnVectors=true before the initial GetParentTask fixes the
issue.

If an editor is already running then the crash doesn't occur
because it responds to the Message_TW_NewTask, so Wimp_StartTask
is never called. If the -ctrl switch is absent then the control
characters from Wimp_StartTask are ignored by MyWrchV. If
-display is absent then it never calls GetParentTask during
the start-up code.

Admin:
Tested standalone on RPCEmu and RPi3B+

Version 0.83. Tagged as 'TaskWindow-0_83'