Sound2: Out of range SWI fixes
When scheduling a SWI, ensure the X bit is set when despatched.
This used to be the case: the version in RISC OS 3.60 forced the X bit, but this was mis-translated to XOS_CallASWI for StrongARM with the X bit clear in R10.
Fixes https://www.riscosopen.org/forum/forums/4/topics/6760

Report an error for attempts to call out of range SWIs from our SWI jump table.

Delete unused Help/Syntax.

Version 1.34. Tagged as 'Sound2-1_34'