RISC_OSLib: Add variants of _swi and _swix that take a va_list
These are typically useful for wrapper functions that want to do something
before and/or after a SWI call, and which therefore need to pass a pointer
to their own variadic arguments into the guts of _swi or _swix.

The pairs of functions _swi and _vswi, and _swix and _vswix, are almost
identical, differing only in argument marshalling.

There being only two additional entries, I didn't think it was worth
introducing a whole new library chunk. Instead, I've tacked them onto the
end of chunk 2 (the basic clib functions).

Note that this means that any applications that use these new entries will
need to take responsibility for RMEnsuring this version of the
SharedCLibrary module and softloading a newer one if necessary.