SprExtend: [540] Fix buffer overflow in ExpandMaskSizeToR0
A mistake in the arguments passed to move_memory_up resulted in the code
moving twice as much data as needed, potentially causing it to write off
the end of the sprite area.

Fix the bad argument, and add a couple of extra comments.

Fixes https://www.riscosopen.org/tracker/tickets/540

Version 1.86. Tagged as 'SprExtend-1_86'