BootCmds: Wildcarded path fix
In BootCmds-1_52 an error was added if the application being AddApp'd doesn't
exist. However, that didn't take into account that AddApp can be used with a
wildcard, and that the wildcard may return no matches. Be quiet in this case.

The new feature to handle paths (eg. myapp:!AppName) could get 'Dir not found'
from OS_GBPB. Fix this by holding the original path internally.

Version 1.53. Tagged as 'BootCmds-1_53'